refactor(types): update types in jobs and node
