+++ src/jobs/types.ts
@@ -1,4 +1,5 @@
-import { Network, type Subset } from "../util";
+import { type Subset } from "../utils";
+import { Network } from "../utils/network";
 
 export type JobNetwork = Subset<Network, "ethereum" | "polygon" | "mumbai">;
 export const JobNetwoks: JobNetwork[] = ["ethereum", "polygon", "mumbai"];
